requirements.txt
Flask
gunicorn
Werkzeug
requests
python-dotenv
google-generativeai